Added all notebooks used in this paper
Each individual file contains all necessary code for the execution of the simulations with the CNN architecture, RPs and time series (TS) data proposed in the methodology section of this paper. They also include all results from training and validation stages.

Data for both the RPs and TS are generated independently from these notebooks, with the respective code included in the previous folder of this repository.